extract kubernetes resource parsers (pod, deployment and metadata) to a separate package
