+++ comp/core/workloadmeta/collectors/internal/kubeapiserver/deployments.go
@@ -10,10 +10,6 @@ package kubeapiserver
 
 import (
 	"context"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"regexp"
-	"strings"
 
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -23,9 +19,9 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
+	kubernetesresourceparsers "github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/util/kubernetes_resource_parsers"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
-	languagedetectionUtil "github.com/DataDog/datadog-agent/pkg/languagedetection/util"
-	ddkube "github.com/DataDog/datadog-agent/pkg/util/kubernetes"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
 // deploymentFilter filters out deployments that can't be used for unified service tagging or process language detection
@@ -59,10 +55,10 @@ func newDeploymentStore(ctx context.Context, wlm workloadmeta.Component, cfg con
 
 func newDeploymentReflectorStore(wlmetaStore workloadmeta.Component, cfg config.Reader) *reflectorStore {
 	annotationsExclude := cfg.GetStringSlice("cluster_agent.kubernetes_resources_collection.deployment_annotations_exclude")
-	parser, err := newdeploymentParser(annotationsExclude)
+	parser, err := kubernetesresourceparsers.NewDeploymentParser(annotationsExclude)
 	if err != nil {
 		_ = log.Errorf("unable to parse all deployment_annotations_exclude: %v, err:", err)
-		parser, _ = newdeploymentParser(nil)
+		parser, _ = kubernetesresourceparsers.NewDeploymentParser(nil)
 	}
 
 	store := &reflectorStore{
@@ -73,71 +69,4 @@ func newDeploymentReflectorStore(wlmetaStore workloadmeta.Component, cfg config.
 	}
 
 	return store
-}
-
-type deploymentParser struct {
-	annotationsFilter []*regexp.Regexp
-	gvr               *schema.GroupVersionResource
-}
-
-func newdeploymentParser(annotationsExclude []string) (objectParser, error) {
-	filters, err := parseFilters(annotationsExclude)
-	if err != nil {
-		return nil, err
-	}
-	return deploymentParser{
-		annotationsFilter: filters,
-		gvr: &schema.GroupVersionResource{
-			Group:    "apps",
-			Version:  "v1",
-			Resource: "deployments",
-		},
-	}, nil
-}
-
-func updateContainerLanguage(cl languagedetectionUtil.ContainersLanguages, container languagedetectionUtil.Container, languages string) {
-	if _, found := cl[container]; !found {
-		cl[container] = make(languagedetectionUtil.LanguageSet)
-	}
-
-	for _, lang := range strings.Split(languages, ",") {
-		cl[container][languagedetectionUtil.Language(strings.TrimSpace(lang))] = struct{}{}
-	}
-}
-
-func (p deploymentParser) Parse(obj interface{}) workloadmeta.Entity {
-	deployment := obj.(*appsv1.Deployment)
-	containerLanguages := make(languagedetectionUtil.ContainersLanguages)
-
-	for annotation, languages := range deployment.Annotations {
-
-		containerName, isInitContainer := languagedetectionUtil.ExtractContainerFromAnnotationKey(annotation)
-		if containerName != "" && languages != "" {
-
-			updateContainerLanguage(
-				containerLanguages,
-				languagedetectionUtil.Container{
-					Name: containerName,
-					Init: isInitContainer,
-				},
-				languages)
-		}
-	}
-
-	return &workloadmeta.KubernetesDeployment{
-		EntityID: workloadmeta.EntityID{
-			Kind: workloadmeta.KindKubernetesDeployment,
-			ID:   deployment.Namespace + "/" + deployment.Name, // we use the namespace/name as id to make it easier for the admission controller to retrieve the corresponding deployment
-		},
-		EntityMeta: workloadmeta.EntityMeta{
-			Name:        deployment.Name,
-			Namespace:   deployment.Namespace,
-			Labels:      deployment.Labels,
-			Annotations: filterMapStringKey(deployment.Annotations, p.annotationsFilter),
-		},
-		Env:                 deployment.Labels[ddkube.EnvTagLabelKey],
-		Service:             deployment.Labels[ddkube.ServiceTagLabelKey],
-		Version:             deployment.Labels[ddkube.VersionTagLabelKey],
-		InjectableLanguages: containerLanguages,
-	}
 }
