+++ comp/core/workloadmeta/collectors/internal/kubeapiserver/pod.go
@@ -9,7 +9,6 @@ package kubeapiserver
 
 import (
 	"context"
-	"regexp"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -19,6 +18,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
+	kubernetesresourceparsers "github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/util/kubernetes_resource_parsers"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -47,89 +47,15 @@ func newPodStore(ctx context.Context, wlm workloadmeta.Component, config config.
 
 func newPodReflectorStore(wlmetaStore workloadmeta.Component, config config.Reader) *reflectorStore {
 	annotationsExclude := config.GetStringSlice("cluster_agent.kubernetes_resources_collection.pod_annotations_exclude")
-	parser, err := newPodParser(annotationsExclude)
+	parser, err := kubernetesresourceparsers.NewPodParser(annotationsExclude)
 	if err != nil {
 		_ = log.Errorf("unable to parse all pod_annotations_exclude: %v, err:", err)
-		parser, _ = newPodParser(nil)
+		parser, _ = kubernetesresourceparsers.NewPodParser(nil)
 	}
 
 	return &reflectorStore{
 		wlmetaStore: wlmetaStore,
 		seen:        make(map[string]workloadmeta.EntityID),
 		parser:      parser,
-	}
-}
-
-type podParser struct {
-	annotationsFilter []*regexp.Regexp
-}
-
-func newPodParser(annotationsExclude []string) (objectParser, error) {
-	filters, err := parseFilters(annotationsExclude)
-	if err != nil {
-		return nil, err
-	}
-
-	return podParser{annotationsFilter: filters}, nil
-}
-
-func (p podParser) Parse(obj interface{}) workloadmeta.Entity {
-	pod := obj.(*corev1.Pod)
-	owners := make([]workloadmeta.KubernetesPodOwner, 0, len(pod.OwnerReferences))
-	for _, o := range pod.OwnerReferences {
-		owners = append(owners, workloadmeta.KubernetesPodOwner{
-			Kind: o.Kind,
-			Name: o.Name,
-			ID:   string(o.UID),
-		})
-	}
-
-	var ready bool
-	for _, condition := range pod.Status.Conditions {
-		if condition.Type == corev1.PodReady {
-			if condition.Status == corev1.ConditionTrue {
-				ready = true
-			}
-			break
-		}
-	}
-
-	var pvcNames []string
-	for _, volume := range pod.Spec.Volumes {
-		if volume.PersistentVolumeClaim != nil {
-			pvcNames = append(pvcNames, volume.PersistentVolumeClaim.ClaimName)
-		}
-	}
-
-	var rtcName string
-	if pod.Spec.RuntimeClassName != nil {
-		rtcName = *pod.Spec.RuntimeClassName
-	}
-
-	return &workloadmeta.KubernetesPod{
-		EntityID: workloadmeta.EntityID{
-			Kind: workloadmeta.KindKubernetesPod,
-			ID:   string(pod.UID),
-		},
-		EntityMeta: workloadmeta.EntityMeta{
-			Name:        pod.Name,
-			Namespace:   pod.Namespace,
-			Annotations: filterMapStringKey(pod.Annotations, p.annotationsFilter),
-			Labels:      pod.Labels,
-		},
-		Phase:                      string(pod.Status.Phase),
-		Owners:                     owners,
-		PersistentVolumeClaimNames: pvcNames,
-		Ready:                      ready,
-		IP:                         pod.Status.PodIP,
-		PriorityClass:              pod.Spec.PriorityClassName,
-		QOSClass:                   string(pod.Status.QOSClass),
-		RuntimeClass:               rtcName,
-
-		// Containers could be generated by this collector, but
-		// currently it's not to save on memory, since this is supposed
-		// to run in the Cluster Agent, and the total amount of
-		// containers can be quite significant
-		// Containers:                 []workloadmeta.OrchestratorContainer{},
 	}
 }
