+++ comp/core/workloadmeta/collectors/internal/kubeapiserver/reflector_store.go
@@ -16,13 +16,9 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
+	kubernetesresourceparsers "github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/util/kubernetes_resource_parsers"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 )
-
-// objectParser is an interface allowing to plug any object
-type objectParser interface {
-	Parse(obj interface{}) workloadmeta.Entity
-}
 
 // entityUID glue together a WLM Entity and a Kube UID
 type entityUID struct {
@@ -35,7 +31,7 @@ type reflectorStore struct {
 
 	mu     sync.Mutex
 	seen   map[string]workloadmeta.EntityID // needs to be updated only if the object is added
-	parser objectParser
+	parser kubernetesresourceparsers.ObjectParser
 	// hasSynced logic is based on the logic see in FIFO queue (client-go/tools/cache/fifo.go)
 	// Normally `Replace` is called first and then `Add/Update/Delete`.
 	// If `Add/Update/Delete` is called first, triggers hasSynced
