+++ comp/core/workloadmeta/collectors/util/kubernetes_resource_parsers/pod.go
@@ -1,0 +1,91 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build kubeapiserver
+
+package kubernetesresourceparsers
+
+import (
+	"regexp"
+
+	corev1 "k8s.io/api/core/v1"
+
+	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
+)
+
+type podParser struct {
+	annotationsFilter []*regexp.Regexp
+}
+
+// NewPodParser creates and returns a pod parser based on annotations exclusion list
+func NewPodParser(annotationsExclude []string) (ObjectParser, error) {
+	filters, err := parseFilters(annotationsExclude)
+	if err != nil {
+		return nil, err
+	}
+
+	return podParser{annotationsFilter: filters}, nil
+}
+
+func (p podParser) Parse(obj interface{}) workloadmeta.Entity {
+	pod := obj.(*corev1.Pod)
+	owners := make([]workloadmeta.KubernetesPodOwner, 0, len(pod.OwnerReferences))
+	for _, o := range pod.OwnerReferences {
+		owners = append(owners, workloadmeta.KubernetesPodOwner{
+			Kind: o.Kind,
+			Name: o.Name,
+			ID:   string(o.UID),
+		})
+	}
+
+	var ready bool
+	for _, condition := range pod.Status.Conditions {
+		if condition.Type == corev1.PodReady {
+			if condition.Status == corev1.ConditionTrue {
+				ready = true
+			}
+			break
+		}
+	}
+
+	var pvcNames []string
+	for _, volume := range pod.Spec.Volumes {
+		if volume.PersistentVolumeClaim != nil {
+			pvcNames = append(pvcNames, volume.PersistentVolumeClaim.ClaimName)
+		}
+	}
+
+	var rtcName string
+	if pod.Spec.RuntimeClassName != nil {
+		rtcName = *pod.Spec.RuntimeClassName
+	}
+
+	return &workloadmeta.KubernetesPod{
+		EntityID: workloadmeta.EntityID{
+			Kind: workloadmeta.KindKubernetesPod,
+			ID:   string(pod.UID),
+		},
+		EntityMeta: workloadmeta.EntityMeta{
+			Name:        pod.Name,
+			Namespace:   pod.Namespace,
+			Annotations: filterMapStringKey(pod.Annotations, p.annotationsFilter),
+			Labels:      pod.Labels,
+		},
+		Phase:                      string(pod.Status.Phase),
+		Owners:                     owners,
+		PersistentVolumeClaimNames: pvcNames,
+		Ready:                      ready,
+		IP:                         pod.Status.PodIP,
+		PriorityClass:              pod.Spec.PriorityClassName,
+		QOSClass:                   string(pod.Status.QOSClass),
+		RuntimeClass:               rtcName,
+
+		// Containers could be generated by this collector, but
+		// currently it's not to save on memory, since this is supposed
+		// to run in the Cluster Agent, and the total amount of
+		// containers can be quite significant
+		// Containers:                 []workloadmeta.OrchestratorContainer{},
+	}
+}
