+++ comp/core/workloadmeta/collectors/util/kubernetes_resource_parsers/utils.go
@@ -1,0 +1,54 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build kubeapiserver
+
+package kubernetesresourceparsers
+
+import (
+	"fmt"
+	"regexp"
+
+	utilserror "k8s.io/apimachinery/pkg/util/errors"
+)
+
+func parseFilters(annotationsExclude []string) ([]*regexp.Regexp, error) {
+	var parsedFilters []*regexp.Regexp
+	var errors []error
+	for _, exclude := range annotationsExclude {
+		filter, err := filterToRegex(exclude)
+		if err != nil {
+			errors = append(errors, err)
+			continue
+		}
+		parsedFilters = append(parsedFilters, filter)
+	}
+
+	return parsedFilters, utilserror.NewAggregate(errors)
+}
+
+// filterToRegex checks a filter's regex
+func filterToRegex(filter string) (*regexp.Regexp, error) {
+	r, err := regexp.Compile(filter)
+	if err != nil {
+		errormsg := fmt.Errorf("invalid regex '%s': %s", filter, err)
+		return nil, errormsg
+	}
+	return r, nil
+}
+
+func filterMapStringKey(mapInput map[string]string, keyFilters []*regexp.Regexp) map[string]string {
+	for key := range mapInput {
+		for _, filter := range keyFilters {
+			if filter.MatchString(key) {
+				delete(mapInput, key)
+				// we can break now since the key is already excluded.
+				break
+			}
+		}
+	}
+
+	return mapInput
+}
