+++ comp/core/workloadmeta/collectors/util/kubernetes_resource_parsers/utils_test.go
@@ -5,7 +5,7 @@
 
 //go:build kubeapiserver
 
-package kubeapiserver
+package kubernetesresourceparsers
 
 import (
 	"reflect"
